rename: Entity 패키지 계층 일관화
