+++ src/main/java/cart/dao/CartDao.java
@@ -1,6 +1,6 @@
 package cart.dao;
 
-import cart.entity.CartEntity;
+import cart.entity.cart.CartEntity;
 import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
