+++ src/main/java/cart/dto/response/ProductResponseDto.java
@@ -1,7 +1,7 @@
 package cart.dto.response;
 
 import cart.dto.ProductCategoryDto;
-import cart.entity.CategoryEntity;
+import cart.entity.category.CategoryEntity;
 import java.util.List;
 import java.util.stream.Collectors;
