+++ src/main/java/cart/entity/cart/CartEntity.java
@@ -1,4 +1,4 @@
-package cart.entity;
+package cart.entity.cart;
 
 public final class CartEntity {
