+++ src/main/java/cart/entity/category/CategoryEntity.java
@@ -1,4 +1,4 @@
-package cart.entity;
+package cart.entity.category;
 
 import java.util.Objects;
