+++ src/main/java/cart/entity/productcategory/ProductCategoryEntity.java
@@ -1,4 +1,4 @@
-package cart.entity;
+package cart.entity.productcategory;
 
 import java.util.Objects;
