+++ src/main/java/cart/service/ProductService.java
@@ -6,9 +6,9 @@ import cart.dao.ProductDao;
 import cart.dto.ProductCategoryDto;
 import cart.dto.request.ProductRequestDto;
 import cart.dto.response.CategoryResponseDto;
-import cart.entity.CategoryEntity;
-import cart.entity.ProductCategoryEntity;
+import cart.entity.category.CategoryEntity;
 import cart.entity.product.ProductEntity;
+import cart.entity.productcategory.ProductCategoryEntity;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
@@ -83,7 +83,7 @@ public class ProductService {
             productRequestDto.getPrice(),
             productRequestDto.getDescription()
         );
-        
+
         productDao.update(savedProductEntity);
         deleteAllLegacyProductCategory(id);
         registerAllProductCategory(productRequestDto, id);
