+++ src/test/java/cart/dao/CategoryDaoTest.java
@@ -2,7 +2,7 @@ package cart.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import cart.entity.CategoryEntity;
+import cart.entity.category.CategoryEntity;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
