+++ src/test/java/cart/entity/CategoryEntityTest.java
@@ -3,6 +3,7 @@ package cart.entity;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import cart.entity.category.CategoryEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
