+++ src/test/java/cart/entity/ProductCategoryEntityTest.java
@@ -3,6 +3,7 @@ package cart.entity;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import cart.entity.productcategory.ProductCategoryEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
