+++ src/test/java/cart/service/CartServiceTest.java
@@ -8,7 +8,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import cart.dao.CartDao;
-import cart.entity.CartEntity;
+import cart.entity.cart.CartEntity;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
