+++ src/test/java/cart/service/ProductServiceTest.java
@@ -12,9 +12,9 @@ import cart.dao.ProductCategoryDao;
 import cart.dao.ProductDao;
 import cart.dto.ProductCategoryDto;
 import cart.dto.request.ProductRequestDto;
-import cart.entity.CategoryEntity;
-import cart.entity.ProductCategoryEntity;
+import cart.entity.category.CategoryEntity;
 import cart.entity.product.ProductEntity;
+import cart.entity.productcategory.ProductCategoryEntity;
 import java.util.List;
 import java.util.Optional;
 import org.assertj.core.api.Assertions;
